UNET model input precision conversion

+++ stable_diffusion_engine.py
@@ -183,7 +183,7 @@ class StableDiffusionEngine:
             # predict the noise residual
             noise_pred = result(self.unet.infer_new_request({
                 "latent_model_input": latent_model_input,
-                "t": t,
+                "t": np.float64(t),
                 "encoder_hidden_states": text_embeddings
             }))
